Improve mesh generation
Source image is mapped more closely to the screen (reduce size of mesh to better fit).
Use z coordinate to widen the mesh on the bottom. Previously this was done by adjusting the x coordinate, but different x coordinates on top and bottom lead to minor distortion. By increasing z, the bottom is more zoomed in and appears wider.